First version of quiz
Example of quiz working, doesn't have the actual questions implemented yet. Couldn't get the python server to work just yet, but this stores a variable of the number correct and displays it on the page when you click submit for now. There is a bit of styling to make it nicer including a dark background with white text for readability, and the text changes color based on the correctness of the answer.